Debugged and Fixed  AB_PlayerMovement
My player animation blueprint continually failed to get a non-None value for the Player Character, and hence the player states could not be evaluated.

Fixed this issue by changing the debug view of my blueprint editor to the appropriate game-instance such that the player object could be found.

Also fixed a minor error where the player's state was always running.